requirements.txt
# Core dependencies
pyautogui
requests
beautifulsoup4
googlesearch-python
ddgs
duckduckgo-search
chromadb
azure-cognitiveservices-speech
python-dotenv
crawl4ai
torch
torchvision
facenet-pytorch
opencv-python
numpy
PyQt6
pillow
pyqtgraph
psutil

# Web scraping and automation
playwright
lxml

# Async support
asyncio
